make function calls suitable for hardware implementation

+++ hll/hls/murmur64a/hll.h
@@ -45,7 +45,7 @@
 
 uint8_t hllPatLen(ap_uint<8> *data, const uint32_t len, uint64_t *idx);
 uint8_t hllSet(uint8_t *registers, uint64_t idx, uint8_t count);
-uint8_t hllAdd(uint8_t *registers, ap_uint<8> *data, const uint32_t len);
+uint64_t hllAdd(uint8_t *registers, ap_uint<8> *data, const uint32_t len);
 void hllRegHisto(uint8_t *registers, uint32_t *reghisto);
 uint64_t hllCount(uint8_t* registers);
